done lab 5! :D

+++ lab_5/cme435_lab5/rm_files.sh
@@ -1,7 +1,25 @@
 #!/usr/bin/env bash
 
-rm -rf work
-rm *.ucdb
-rm transcript
-rm *.rpt
-rm -rf pdm_cov_html
+if [ -d "code_coverage" ]; then
+	rm -rf code_coverage
+fi
+
+if [ -d "work" ]; then
+	rm -rf work
+fi
+
+if [ -f *.ucdb ]; then
+	rm *.ucdb
+fi
+
+if [ -f transcript ]; then
+	rm transcript
+fi
+
+if [ -f *.rpt ]; then
+	rm *.rpt
+fi
+
+if [ -d "code_coverage/pdm_cov_html" ]; then
+	rm -rf code_coverage/pdm_cov_html
+fi
